requirements.txt
ffmpeg_python==0.2.0
instagrapi==1.17.8
loguru==0.7.0
mediapipe==0.9.3.0
opencv_contrib_python==4.5.5.62
opencv_python==4.7.0.72
python-dotenv==1.0.0
torch==2.0.1
